META-3040 Filter via purpose/persona in Indexsearch

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/AtlasRelationshipStoreV2.java
@@ -69,6 +69,7 @@ import java.util.Set;
 import java.util.UUID;
 
 import static org.apache.atlas.AtlasConfiguration.NOTIFICATION_RELATIONSHIPS_ENABLED;
+import static org.apache.atlas.accesscontrol.AccessControlUtil.ensureNonAccessControlRelType;
 import static org.apache.atlas.model.instance.AtlasEntity.Status.DELETED;
 import static org.apache.atlas.model.typedef.AtlasRelationshipDef.PropagateTags.BOTH;
 import static org.apache.atlas.model.typedef.AtlasRelationshipDef.PropagateTags.NONE;
@@ -308,7 +309,7 @@ public class AtlasRelationshipStoreV2 implements AtlasRelationshipStore {
             if (getState(edge) == DELETED) {
                 throw new AtlasBaseException(AtlasErrorCode.RELATIONSHIP_ALREADY_DELETED, guid);
             }
-
+            ensureNonAccessControlRelType(getTypeName(edge));
             edgesToDelete.add(edge);
             deletedRelationships.add(entityRetriever.mapEdgeToAtlasRelationship(edge));
         }
@@ -361,6 +362,7 @@ public class AtlasRelationshipStoreV2 implements AtlasRelationshipStore {
         if (getState(edge) == DELETED) {
             throw new AtlasBaseException(AtlasErrorCode.RELATIONSHIP_ALREADY_DELETED, guid);
         }
+        ensureNonAccessControlRelType(getTypeName(edge));
         deleteDelegate.getHandler().resetHasLineageOnInputOutputDelete(Collections.singleton(edge), null);
         deleteDelegate.getHandler().deleteRelationships(Collections.singleton(edge), forceDelete);
